kernel - remove bounds on buffer cache nbuf count for 64-bit
* Remove arbitrary 1GB buffer cache limitation

* Adjusted numerous 'int' fields to 'long'.  Even though nbuf is not
  likely to exceed 2 billion buffers, byte calculations using the
  variable began overflowing so just convert that and various other
  variables to long.

* Make sure we don't blow-out the temporary valloc() space in early boot
  due to nbufs being too large.

* Unbound 'kern.nbuf' specifications in /boot/loader.conf as well.

+++ sys/sys/buf.h
@@ -376,7 +376,7 @@ struct cluster_save {
 #define B_SYNC		0x02		/* Do all allocations synchronously. */
 
 #ifdef _KERNEL
-extern int	nbuf;			/* The number of buffer headers */
+extern long	nbuf;			/* The number of buffer headers */
 extern long	maxswzone;		/* Max KVA for swap structures */
 extern long	maxbcache;		/* Max KVA for buffer cache */
 extern long	hidirtybufspace;
@@ -385,7 +385,7 @@ extern struct buf *buf;			/* The buffer headers. */
 extern char	*buffers;		/* The buffer contents. */
 extern int	bufpages;		/* Number of memory pages in the buffer pool. */
 extern struct	buf *swbuf;		/* Swap I/O buffer headers. */
-extern int	nswbuf;			/* Number of swap I/O buffer headers. */
+extern long	nswbuf;			/* Number of swap I/O buffer headers. */
 extern int	bioq_reorder_burst_interval;
 extern int	bioq_reorder_burst_bytes;
 extern int	bioq_reorder_minor_interval;
